shakedex: move LISTING_STATUS to constants file
