+++ app/components/ShakedexDeprecated/index.js
@@ -7,7 +7,8 @@ import GenerateListingModal from '../../pages/Exchange/GenerateListingModal';
 import Alert from '../Alert';
 import MiniModal from '../Modal/MiniModal';
 import { HeaderItem, HeaderRow, Table, TableItem, TableRow } from '../Table';
-import { cancelExchangeLock, finalizeCancelExchangeLock, getExchangeListings, LISTING_STATUS } from "../../ducks/exchange";
+import { cancelExchangeLock, finalizeCancelExchangeLock, getExchangeListings } from "../../ducks/exchange";
+import { LISTING_STATUS } from '../../constants/exchange';
 import { formatName } from '../../utils/nameHelpers';
 import { listingStatusToI18nKey } from '../../utils/shakedex';
 import { I18nContext } from '../../utils/i18n';
