+++ app/constants/exchange.js
@@ -1,0 +1,14 @@
+export const LISTING_STATUS = {
+  NOT_FOUND: 'NOT_FOUND',
+  TRANSFER_CONFIRMING: 'TRANSFER_CONFIRMING',
+  TRANSFER_CONFIRMED: 'TRANSFER_CONFIRMED',
+  TRANSFER_CONFIRMED_LOCKUP: 'TRANSFER_CONFIRMED_LOCKUP',
+  FINALIZE_CONFIRMING: 'FINALIZE_CONFIRMING',
+  FINALIZE_CONFIRMED: 'FINALIZE_CONFIRMED',
+  ACTIVE: 'ACTIVE',
+  SOLD: 'SOLD',
+  CANCEL_CONFIRMING: 'CANCEL_CONFIRMING',
+  CANCEL_CONFIRMED: 'CANCEL_CONFIRMED',
+  FINALIZE_CANCEL_CONFIRMING: 'FINALIZE_CANCEL_CONFIRMING',
+  FINALIZE_CANCEL_CONFIRMED: 'FINALIZE_CANCEL_CONFIRMED',
+};
